# Primary caching 5: 2D & 3D point clouds