--- a/crates/re_space_view_spatial/Cargo.toml
+++ b/crates/re_space_view_spatial/Cargo.toml
@@ -24,6 +24,7 @@ re_format.workspace = true
 re_log_types.workspace = true
 re_log.workspace = true
 re_query.workspace = true
+re_query_cache.workspace = true
 re_renderer = { workspace = true, features = ["import-gltf", "import-obj"] }
 re_types = { workspace = true, features = ["ecolor", "glam", "image"] }
 re_tracing.workspace = true