--- a/crates/re_space_view_spatial/src/lib.rs
+++ b/crates/re_space_view_spatial/src/lib.rs
@@ -22,7 +22,7 @@ pub use space_view_2d::SpatialSpaceView2D;
 pub use space_view_3d::SpatialSpaceView3D;
 
 #[doc(hidden)] // Public for benchmarks
-pub use visualizers::LoadedPoints;
+pub use visualizers::{LoadedPoints, Points3DComponentData};
 
 // ---